django-jwt-mysql/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Markdown==3.3.4
mysqlclient==2.0.3
pydot==1.4.2
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
